# docs/swagger.yaml
basePath: /api
definitions:
  controllers.LoginInput:
    properties:
      npp:
        type: string
      password:
        type: string
    required:
    - npp
    - password
    type: object
...
  request.ReviewApprovalRequest:
    properties:
      action:
        description: 'Action: "APPROVE" atau "REVISION"'
        type: string
      notes:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for Backend Idle (equipment/idle management system)
  title: Backend Idle API
  version: "1.0"
paths:
  /approvals:
    get:
      description: Mengambil daftar approval request, dengan filter opsional status
        dan request_type
      parameters:
      - description: 'Filter status approval, contoh: PENDING'
        in: query
        name: status
        type: string
      - description: 'Filter tipe request, contoh: EQUIPMENT_REGISTRATION'
        in: query
        name: request_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List approval requests
      tags:
      - approvals
  /approvals/{id}:
    get:
      description: Mengambil detail satu approval request beserta tahapan (steps)
      parameters:
      - description: Approval Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get approval detail
      tags:
      - approvals
  /approvals/{id}/review:
    patch:
      consumes:
      - application/json
      description: Memproses keputusan Manajer Rendal terhadap approval request (APPROVE
        atau REVISION). Jika APPROVE, equipment terkait otomatis berubah menjadi IDLE.
      parameters:
      - description: Approval Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ReviewApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Review approval request
      tags:
      - approvals
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login menggunakan NPP dan password, mengembalikan JWT token beserta
        data user
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Ketik "Bearer" diikuti spasi dan token JWT. Contoh: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"